backend/requirements.txt
greenlet==3.1.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiomysql==0.2.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.28.1
